Create SMART AND INTELLIGENT POWERED IRON BOX
Iron boxes are used in our day to day life .In these iron boxes temperature is controlled mechanically. We all are leading a busy life so we forgot to adjust the temperature before ironing the clothes .It will damage the clothes. To overcome this issue, we are presenting a new idea-"SMART AND INTELLIGENT POWERED IRON. BOX". In this, the temperature controlling is fully automatic Temperature controlling is done with two sensors ,one is Fabric detecting sensor Which senses &detect different types of fabrics such as cotton ,polyester,nylon etc. and the second sensor is a temperature controlling sensor which automatically set the temperature of Iron box according to different fabrics .When we are ironing different fabrics ,the temperature will be setted up immediately(fastly).Firstly the temperature becomes a moderate value then it changes to the required temperature needed for the ironing of fabrics. AI based image processing technique is used in the fabric detecting sensor. Selecting the correct setting on an iron box can make the difference between a good job and a disaster.
It is a frequent problem that our clothes get burnt while ironing. This mainly happens when the iron box is left idle when attending any phone calls or engaging in any other activities. Ironing works by loosening the bond between long chain polymer molecules in the fibers of the material. SMART AND INTELLIGENT POWERED IRON BOX preventing fabric damage.It design that uses an innovative framework and motorized mechanisms to effectively iron various clothes by minimizing the difficulties in the task of ironing.